move country select into own file
